% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.month.R
\name{months_begin_ymd}
\alias{months_begin_ymd}
\title{Calculate difference (in months) between 2 dates (ymd format). it assumes that you are interestedin the begining of each month.}
\usage{
months_begin_ymd(begin, end)
}
\arguments{
\item{begin}{A date, in format ymd}

\item{end}{A date, in format ymd}
}
\value{
difference (in months) between begin and end
}
\description{
Calculate difference (in months) between 2 dates (ymd format). it assumes that you are interestedin the begining of each month.
}
\examples{
months_begin_ymd("201161211", "20161215")
}

