% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.month.R
\name{Months_dmy}
\alias{Months_dmy}
\title{Calculate difference (in months) between 2 dates}
\usage{
Months_dmy(d1, d2)
}
\arguments{
\item{d1}{A date, in format dmy}

\item{d2}{A date, in format dmy}
}
\value{
difference (in months) between d1 and d2
}
\description{
Calculate difference (in months) between 2 dates
}
\examples{
Months_dmy("11122016", "12152016")
}

