% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.month.R
\name{Months_mdy}
\alias{Months_mdy}
\title{Calculate difference (in months) between 2 dates}
\usage{
Months_mdy(d1, d2)
}
\arguments{
\item{d1}{A date, in format mdy}

\item{d2}{A date, in format mdy}
}
\value{
difference (in months) between d1 and d2
}
\description{
Calculate difference (in months) between 2 dates
}
\examples{
Months_mdy("12112016", "15122016")
}

