% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.month.R
\name{Months_ymd}
\alias{Months_ymd}
\title{Calculate difference (in months) between 2 dates}
\usage{
Months_ymd(d1, d2)
}
\arguments{
\item{d1}{A date, in format ymd}

\item{d2}{A date, in format ymd}
}
\value{
difference (in months) between d1 and d2
}
\description{
Calculate difference (in months) between 2 dates
}
\examples{
Months_ymd("201161211", "20161215")
}

