% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.month.R
\name{mos_ymd}
\alias{mos_ymd}
\title{Calculate difference (in months) between 2 dates. it assumes that you are interestedin the REAL difference between each date}
\usage{
mos_ymd(begin, end)
}
\arguments{
\item{begin}{A date, in format ymd}

\item{end}{A date, in format ymd}
}
\value{
difference (in months) between begin and end
}
\description{
Calculate difference (in months) between 2 dates. it assumes that you are interestedin the REAL difference between each date
}
\examples{
mos_ymd("201161211", "20161215")
}

