% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.month.R
\name{months_mdy}
\alias{months_mdy}
\title{Calculate difference (in months) between 2 dates (mdy format)}
\usage{
months_mdy(d1, d2)
}
\arguments{
\item{d1}{A date, in format mdy}

\item{d2}{A date, in format mdy}
}
\value{
difference (in months) between d1 and d2
}
\description{
Calculate difference (in months) between 2 dates (mdy format)
}
\examples{
months_mdy("12112016", "15122016")
}

